+++ src/main/java/com/fmt/educonnect/datasource/repositories/AlunoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlunoRepository extends JpaRepository<AlunoEntity, Integer> {
+public interface AlunoRepository extends JpaRepository<AlunoEntity, Long> {
 }
